cleaning up module name etc
